Back-port pushmi from C++20+modules to standard C++14. (#16)
* builds (with warnings) and runs correctly with -std=c++14 -fconcepts

* emulate if constexpr on C++14

* Use an And<> concept instead of fold expressions

* replace use of C++20-style lambdas with C++14-style generic lambdas with a constrain<XXX>(yyy) wrapper for adding constraints

* adding property_set and updating Concepts

properties are an extension point so they can be in a class or applied by external trait specialization.

public:
  using properties - property_set<receiver_tag, single_tag, flow_tag, time_tag>;

Each property is also an extension point so that the category can be inside a property type or applied externally by specialization.

property_set itself is a meta type that should be built on a meta ;) library. more work is needed to make manipulation simpler.

concepts are added to query the properties for an object.

PropertyQuery<>
PropertySetQuery<>
PropertyCategoryQuery<>
PropertySetCategoryQuery<>

* split out properties.h

this revamps the properties to use traits, not concepts

updates all the concepts to use the new properties

updates the tags to be both tags and traits

fixes Many to refine None not Single

got rid of the T...

* restore existing Concept naming
port to C++14

* reimplement property_from_category_t

* rebased properties usage on to latest cpp14-support

this compiles and runs all tests. I had to comment out two requirements in tap() to get here.

* cmake builds a single header include/pushmi.h

the primary benefit is for use in godbolt.

* use singleheader

* add concepts emulation layer, begin to convert over to using concepts emulation

* reorder includes in buildSingleHeader

* adding executors-impl as submodule

* add mi as alias

remove v:: later

* add for_each example

wraps static_thread_pool
defines bulk and uses bulk in for_each

* add reduce example

* fix minor compile problems in examples

* race on pending count

I was concerned about this, but for reduce I changed the initial pending count from N to 0 to avoid using std::distance on forward iterators.

this change adds a pending count to the for loop itself so that completion does not occur until the for loop has completed.

* Examples readme (#7)

extending readme to cover for_each and reduce

* Migrate Receiver concept and friends over to concepts emulation

* Futures merge (#8)

* adds composition_2 example
* add subject, share, defer and new example that compares the lazy vs eager options for pushmi

* convert the rest of the concepts to emulation

* Semi example (#10)

* add request_via and composition_4
* remove PUSHMI_TEMPLATE workaround from subject

* add and use via_cast (#11)

* convert some real requires clauses to fake ones.

* Twoway and then (#12)

* adding futures-impl to external
* add then_execute and twoway_execute

* pushmi builds and runs successfully in standard C++14, without concepts

* add alias for property_from_category_t (#14)

* forgot to build single header (#15)

forgot single header